+++ packages/react-tinacms/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.8](https://github.com/tinacms/tinacms/compare/react-tinacms@0.9.8-alpha.1...react-tinacms@0.9.8) (2020-02-03)
+
+**Note:** Version bump only for package react-tinacms
+
+
+
+
+
+## [0.9.8-alpha.1](https://github.com/tinacms/tinacms/compare/react-tinacms@0.9.7...react-tinacms@0.9.8-alpha.1) (2020-02-03)
+
+**Note:** Version bump only for package react-tinacms
+
+
+
+
+
 ## [0.9.8-alpha.0](https://github.com/tinacms/tinacms/compare/react-tinacms@0.9.7...react-tinacms@0.9.8-alpha.0) (2020-01-29)
 
 **Note:** Version bump only for package react-tinacms